Move mind map view into Home.vue
